move util functions to /utils
