+++ src/modules/olMap/components/OlMap.js
@@ -8,7 +8,7 @@ import { defaults as defaultInteractions, PinchRotate } from 'ol/interaction';
 import { removeLayer } from '../../../store/layers/layers.action';
 import { changeLiveRotation, changeZoomCenterAndRotation } from '../../../store/position/position.action';
 import { $injector } from '../../../injection';
-import { updateOlLayer, toOlLayerFromHandler, registerLongPressListener, getLayerById } from '../olMapUtils';
+import { updateOlLayer, toOlLayerFromHandler, registerLongPressListener, getLayerById } from '../utils/olMapUtils';
 import { setBeingDragged, setClick, setContextClick, setPointerMove } from '../../../store/pointer/pointer.action';
 import { setBeingMoved, setMoveEnd, setMoveStart } from '../../../store/map/map.action';
 import VectorSource from 'ol/source/Vector';
