+++ src/modules/olMap/handler/measure/MeasurementOverlayStyle.js
@@ -1,7 +1,7 @@
 import { $injector } from '../../../../injection';
 import { OverlayStyle } from '../../OverlayStyle';
 import { MeasurementOverlayTypes } from './MeasurementOverlay';
-import { getAzimuth, getLineString, getPartitionDelta } from '../../olGeometryUtils';
+import { getAzimuth, getLineString, getPartitionDelta } from '../../utils/olGeometryUtils';
 import Overlay from 'ol/Overlay';
 import { LineString, Polygon } from 'ol/geom';
 import MapBrowserEventType from 'ol/MapBrowserEventType';
