+++ src/modules/olMap/utils/olGeometryUtils.js
