+++ src/modules/olMap/utils/olMapUtils.js
