+++ test/modules/olMap/tooltip/HelpTooltip.test.js
@@ -1,7 +1,7 @@
 import { Overlay } from 'ol';
 import { $injector } from '../../../../src/injection';
 import { HelpTooltip } from '../../../../src/modules/olMap/tooltip/HelpTooltip';
-import { InteractionSnapType, InteractionStateType } from '../../../../src/modules/olMap/olInteractionUtils';
+import { InteractionSnapType, InteractionStateType } from '../../../../src/modules/olMap/utils/olInteractionUtils';
 import { TestUtils } from '../../../test-utils.js';
 import { provide as measureProvide } from '../../../../src/modules/olMap/handler/measure/tooltipMessage.provider';
